# Trailing commas should be removed when last item in the object is deleted