**Description**: 
Right now when last item in object is deleted, trailing commas remain and make json invalid.

For example: 
Given such a json: 
```{ "a": "1", "b": "2"}```

If we delete `b` key then the result will be: 
```{ "a": "1",}```

This makes json invalid. In this PR I've fixed it.

**Benchmark before change**:
```
BenchmarkJsonParserLarge-4                        100000             69483 ns/op               0 B/op          0 allocs/op
BenchmarkJsonParserMedium-4                      1000000             11553 ns/op               0 B/op          0 allocs/op
BenchmarkJsonParserDeleteMedium-4                 500000             11788 ns/op               0 B/op          0 allocs/op
BenchmarkJsonParserEachKeyManualMedium-4         1000000              7436 ns/op             112 B/op          2 allocs/op
BenchmarkJsonParserEachKeyStructMedium-4         1000000              7962 ns/op             672 B/op         12 allocs/op
BenchmarkJsonParserObjectEachStructMedium-4      1000000             11725 ns/op             624 B/op         11 allocs/op
BenchmarkJsonParserSmall-4                      10000000              1535 ns/op               0 B/op          0 allocs/op
BenchmarkJsonParserEachKeyManualSmall-4         10000000               869 ns/op              80 B/op          2 allocs/op
BenchmarkJsonParserEachKeyStructSmall-4         10000000              1074 ns/op             144 B/op          4 allocs/op
BenchmarkJsonParserObjectEachStructSmall-4      10000000               921 ns/op             128 B/op          3 allocs/op
BenchmarkJsonParserSetSmall-4                    5000000              1534 ns/op             816 B/op          5 allocs/op
BenchmarkJsonParserDelSmall-4                    3000000              2035 ns/op               0 B/op          0 allocs/op
```

**Benchmark after change**:

```
BenchmarkJsonParserLarge-4                        100000             65019 ns/op               0 B/op          0 allocs/op
BenchmarkJsonParserMedium-4                      1000000             11508 ns/op               0 B/op          0 allocs/op
BenchmarkJsonParserDeleteMedium-4                 500000             13632 ns/op               0 B/op          0 allocs/op
BenchmarkJsonParserEachKeyManualMedium-4         1000000              6767 ns/op             112 B/op          2 allocs/op
BenchmarkJsonParserEachKeyStructMedium-4         1000000              7821 ns/op             672 B/op         12 allocs/op
BenchmarkJsonParserObjectEachStructMedium-4      1000000             11342 ns/op             624 B/op         11 allocs/op
BenchmarkJsonParserSmall-4                      10000000              1125 ns/op               0 B/op          0 allocs/op
BenchmarkJsonParserEachKeyManualSmall-4         10000000               883 ns/op              80 B/op          2 allocs/op
BenchmarkJsonParserEachKeyStructSmall-4         10000000              1072 ns/op             144 B/op          4 allocs/op
BenchmarkJsonParserObjectEachStructSmall-4      10000000               852 ns/op             128 B/op          3 allocs/op
BenchmarkJsonParserSetSmall-4                    5000000              1601 ns/op             816 B/op          5 allocs/op
BenchmarkJsonParserDelSmall-4                    5000000              1986 ns/op               0 B/op          0 allocs/op
```


For running benchmarks use:
```
go test -test.benchmem -bench JsonParser ./benchmark/ -benchtime 5s -v
# OR
make bench (runs inside docker)
```